tweak template
now that I'm using rust-analyzer with vscode regularly, I've realized
that I _really_ lean on the autocomplete to write code quickly. alas, it
doesn't work in a macro context...

while the macro approach did indeed save the need to copy-paste for q2,
the extra overhead of that pales in comparison to the speed loss of not
having autocomplete.